perf: use float-precision math functions in cacao\n\nReplace double-precision math functions with float-precision\nvariants across 15 cacao source files where the operands and\nresults are float type:\n\n- sqrt() -> sqrtf(), pow() -> powf(), sin() -> sinf(),\n  cos() -> cosf(), exp() -> expf(), fabs() -> fabsf(),\n  floor() -> floorf(), ceil() -> ceilf(), log10() -> log10f()\n- Double literal constants -> float literals (e.g. 0.5 -> 0.5f)\n  in float arithmetic contexts\n- Guard diagnostic printf with UNLIKELY(data.core.Debug > 0)\n  in DMturbulence.c hot path\n\nPreserves double-precision calls where double variables are\nintentionally used for numerical stability (accumulators,\nintermediate calculations).\n\nModified files:\n  AOloopControl_DM/{DMturbulence,pokerndmodes}.c\n  AOloopControl/{modalstatsTUI,modalfilter,modalfilter_test,modalCTRL_stats}.c\n  AOloopControl_PredictiveControl/setPFsimpleAve.c\n  AOloopControl_IOtools/WFSmap.c\n  AOloopControl_acquireCalib/measure_linear_resp.c\n  AOloopControl_perfTest/{mlat,compRMsensitivity,AOloopControl_perfTest}.c\n  computeCalib/{AOloopControl_computeCalib_dm,maskextrapolate}.c\n  pyramidWFStools/pyWFSgridmatch.c

+++ AOloopControl/modalCTRL_stats.c
@@ -537,12 +537,12 @@ static errno_t compute_function()
                             "%7.3f   [nm]  --> %5.3f\n",
                             block,
                             block_cnt[block],
-                            1000.0 * sqrt(block_WFSrms2[block]),
-                            1000.0 * sqrt(block_WFSmrms2[block]),
-                            1000.0 * sqrt(block_WFSmqrms2[block]),
-                            1000.0 * sqrt(block_DMrms2[block]),
-                            1000.0 * sqrt(block_OLrms2[block]),
-                            sqrt(block_WFSrms2[block]) / sqrt(block_OLrms2[block]));
+                            1000.0 * sqrtf(block_WFSrms2[block]),
+                            1000.0 * sqrtf(block_WFSmrms2[block]),
+                            1000.0 * sqrtf(block_WFSmqrms2[block]),
+                            1000.0 * sqrtf(block_DMrms2[block]),
+                            1000.0 * sqrtf(block_OLrms2[block]),
+                            sqrtf(block_WFSrms2[block]) / sqrtf(block_OLrms2[block]));
 
                         char ffname[STRINGMAXLEN_FULLFILENAME];
                         WRITE_FULLFILENAME(ffname, "AOmodalstat.dat");
@@ -551,12 +551,12 @@ static errno_t compute_function()
                                 "%5ld  %02d   %7.3f %7.3f %7.3f %7.3f  %5.3f\n",
                                 processinfo->loopcnt,
                                 block,
-                                1000.0 * sqrt(block_WFSrms2[block]),
-                                1000.0 * sqrt(block_WFSmqrms2[block]),
-                                1000.0 * sqrt(block_DMrms2[block]),
-                                1000.0 * sqrt(block_OLrms2[block]),
-                                sqrt(block_WFSrms2[block]) /
-                                sqrt(block_OLrms2[block]));
+                                1000.0 * sqrtf(block_WFSrms2[block]),
+                                1000.0 * sqrtf(block_WFSmqrms2[block]),
+                                1000.0 * sqrtf(block_DMrms2[block]),
+                                1000.0 * sqrtf(block_OLrms2[block]),
+                                sqrtf(block_WFSrms2[block]) /
+                                sqrtf(block_OLrms2[block]));
                         fclose(fp);
                     }
                 }
